+++ Tests/BitArrayTests/BitArrayTests.swift
@@ -2,7 +2,7 @@
 //  File.swift
 //  
 //
-//  Created by Mahanaz Atiqullah on 6/9/21.
+//  Created by Mahanaz Atiqullah on 7/27/21.
 //
 
 import XCTest
@@ -11,336 +11,10 @@ import CollectionsTestSupport
 
 final class BitArrayTest: CollectionTestCase {
   
-  // DEFINITELY needs improvement lol. I can imagine this looking like a terrible butcher to the experienced eye.
-  let limit = 100 // if change limit, change getStorageCount and getExcessCount tests as well
+  let sizes: [Int] = _getSizes(BitArray.UNIT.bitWidth)
   
-  func testArrayLitInit() {
-    
-    let array = [true, false, true, false, true, true, false, true, false, true]
-    
-    let bitArray = BitArray(arrayLiteral: true, false, true, false, true, true, false, true, false, true)
-    let bitArray2 = BitArray(array)
-    
-    XCTAssertEqual(bitArray.storage, [181, 2])
-    XCTAssertEqual(bitArray.excess, 2)
-    XCTAssertEqual(bitArray.endIndex, bitArray.count)
-    XCTAssertEqual(bitArray.endIndex, 10)
-    
-    XCTAssertEqual(bitArray2.storage, [181, 2])
-    XCTAssertEqual(bitArray2.excess, 2)
-    XCTAssertEqual(bitArray2.endIndex, bitArray.count)
-    XCTAssertEqual(bitArray2.endIndex, 10)
-    
-    XCTAssertEqual(bitArray2, bitArray)
-    
+  func testAppend() {
+    withSomeUsefulBitArrays("bitArray", ofSizes: sizes, ofUnitBitWidth: BitArray.UNIT.bitWidth) { layout in
+    }
   }
-  
-  func testRepeatingValInit() {
-    let bitArray1 = BitArray(repeating: true, count: 0)
-    let bitArray2 = BitArray(repeating: true, count: 1)
-    let bitArray3 = BitArray(repeating: true, count: 7)
-    let bitArray4 = BitArray(repeating: true, count: 8)
-    let bitArray5 = BitArray(repeating: true, count: 9)
-    let bitArray6 = BitArray(repeating: true, count: 15)
-    let bitArray7 = BitArray(repeating: true, count: 16)
-    let bitArray8 = BitArray(repeating: true, count: 17)
-    
-    let falseBitArray1 = BitArray(repeating: false, count: 0)
-    let falseBitArray2 = BitArray(repeating: false, count: 1)
-    let falseBitArray3 = BitArray(repeating: false, count: 7)
-    let falseBitArray4 = BitArray(repeating: false, count: 8)
-    let falseBitArray5 = BitArray(repeating: false, count: 9)
-    let falseBitArray6 = BitArray(repeating: false, count: 15)
-    let falseBitArray7 = BitArray(repeating: false, count: 16)
-    let falseBitArray8 = BitArray(repeating: false, count: 17)
-    
-    XCTAssertEqual(bitArray1.storage, [])
-    XCTAssertEqual(bitArray1.count, 0)
-    XCTAssertEqual(bitArray2.storage, [1])
-    XCTAssertEqual(bitArray2.count, 1)
-    XCTAssertEqual(bitArray3.storage, [127])
-    XCTAssertEqual(bitArray3.count, 7)
-    XCTAssertEqual(bitArray4.storage, [BitArray.UNIT.max])
-    XCTAssertEqual(bitArray4.count, 8)
-    XCTAssertEqual(bitArray5.storage, [BitArray.UNIT.max, 1])
-    XCTAssertEqual(bitArray5.count, 9)
-    XCTAssertEqual(bitArray6.storage, [BitArray.UNIT.max, 127])
-    XCTAssertEqual(bitArray6.count, 15)
-    XCTAssertEqual(bitArray7.storage, [BitArray.UNIT.max, BitArray.UNIT.max])
-    XCTAssertEqual(bitArray7.count, 16)
-    XCTAssertEqual(bitArray8.storage, [BitArray.UNIT.max, BitArray.UNIT.max, 1])
-    XCTAssertEqual(bitArray8.count, 17)
-    
-    XCTAssertEqual(falseBitArray1.storage, [])
-    XCTAssertEqual(falseBitArray1.count, 0)
-    XCTAssertEqual(falseBitArray2.storage, [0])
-    XCTAssertEqual(falseBitArray2.count, 1)
-    XCTAssertEqual(falseBitArray3.storage, [0])
-    XCTAssertEqual(falseBitArray3.count, 7)
-    XCTAssertEqual(falseBitArray4.storage, [0])
-    XCTAssertEqual(falseBitArray4.count, 8)
-    XCTAssertEqual(falseBitArray5.storage, [0, 0])
-    XCTAssertEqual(falseBitArray5.count, 9)
-    XCTAssertEqual(falseBitArray6.storage, [0, 0])
-    XCTAssertEqual(falseBitArray6.count, 15)
-    XCTAssertEqual(falseBitArray7.storage, [0, 0])
-    XCTAssertEqual(falseBitArray7.count, 16)
-    XCTAssertEqual(falseBitArray8.storage, [0, 0, 0])
-    XCTAssertEqual(falseBitArray8.count, 17)
-  
-  }
-  
-  func testRemoveAll() {
-    var testArray = BitArray()
-    var value = Bool.random()
-    
-    for _ in 0...limit {
-      testArray.append(value)
-      value = Bool.random()
-    }
-    
-    //testArray.removeAll()
-    
-    //XCTAssertEqual(testArray.count, 0)
-    //XCTAssertEqual(testArray.endIndex, 0)
-    //XCTAssertEqual(testArray.startIndex, 0)
-    //XCTAssertEqual(testArray.getStorageCount(), 0)
-    //XCTAssertEqual(testArray.getExcessCount(), 0)
-  }
-  
-  func testSubscriptReturnOnly() {
-    var testArray = BitArray()
-    var value = Bool.random()
-    
-    
-    for _ in 0...limit {
-      testArray.append(value)
-      value = Bool.random()
-    }
-    
-    XCTAssertEqual(testArray.count, limit+1)
-    XCTAssertEqual(testArray.endIndex, limit+1)
-    XCTAssertEqual(testArray.startIndex, 0)
-    //XCTAssertEqual(testArray.getStorageCount(), 13)
-    //XCTAssertEqual(testArray.getExcessCount(), 5)
-  }
-  
-  func testSubscriptSetOnly() {
-    
-    var testArray = BitArray()
-    var value = Bool.random()
-    
-    
-    for _ in 0...limit {
-      testArray.append(value)
-      value = Bool.random()
-    }
-    
-    for i in 0..<testArray.endIndex {
-      let setVal = Bool.random()
-      let copy = testArray
-      testArray[i] = setVal
-      XCTAssertEqual(testArray[i], setVal)
-      XCTAssertNotEqual(testArray[i], !setVal)
-      
-      testArray[i] = !setVal
-      XCTAssertEqual(testArray[i], !setVal)
-      XCTAssertNotEqual(testArray[i], setVal)
-      
-      // testing that previous values are unaffected
-      for j in 0..<i {
-        XCTAssertEqual(copy[j], testArray[j])
-      }
-    }
-    
-  }
-  
-  func testformBitwiseOR(){
-    var testArray = BitArray()
-    var testArray2 = BitArray()
-    
-    for i in 0...limit {
-      testArray.append((i%2 == 0))
-    }
-    
-    for i in 0...limit {
-      testArray2.append((i%2 != 0))
-    }
-    
-    
-    testArray.formBitwiseOr(with: testArray2)
-    
-    var resultStorage: [UInt8] = []
-    
-    for _ in 0..<12 {
-      resultStorage.append(BitArray.UNIT.max)
-    }
-    resultStorage.append(31)
-    
-    XCTAssertEqual(testArray.storage.count, 13)
-    XCTAssertEqual(testArray.storage, resultStorage)
-    
-  }
-  
-  func testformBitwiseAND(){
-    var testArray = BitArray()
-    var testArray2 = BitArray()
-    
-    for i in 0...limit {
-      testArray.append((i%2 == 0))
-    }
-    
-    for i in 0...limit {
-      testArray2.append((i%2 != 0))
-    }
-    
-    
-    testArray.formBitwiseAnd(with: testArray2)
-    
-    var resultStorage: [UInt8] = []
-    
-    for _ in 0..<12 {
-      resultStorage.append(0)
-    }
-    resultStorage.append(0)
-    
-    XCTAssertEqual(testArray.storage.count, 13)
-    XCTAssertEqual(testArray.storage, resultStorage)
-    
-  }
-  
-  func testformBitwiseXOR(){
-    var testArray = BitArray()
-    var testArray2 = BitArray()
-    
-    for i in 0...limit {
-      testArray.append((i%2 == 0))
-    }
-    
-    for i in 0...limit {
-      testArray2.append((i%2 != 0))
-    }
-    
-    
-    testArray.formBitwiseXor(with: testArray2)
-    
-    var resultStorage: [UInt8] = []
-    
-    for _ in 0..<12 {
-      resultStorage.append(BitArray.UNIT.max)
-    }
-    resultStorage.append(31)
-    
-    XCTAssertEqual(testArray.storage.count, 13)
-    XCTAssertEqual(testArray.storage, resultStorage)
-    
-  }
-  
-  func testBitwiseOR(){
-    var testArray = BitArray()
-    var testArray2 = BitArray()
-    
-    for i in 0...limit {
-      testArray.append((i%2 == 0))
-    }
-    
-    for i in 0...limit {
-      testArray2.append((i%2 != 0))
-    }
-    
-    
-    let newArr = testArray.bitwiseOr(with: testArray2)
-    
-    var resultStorage: [UInt8] = []
-    
-    for _ in 0..<12 {
-      resultStorage.append(BitArray.UNIT.max)
-    }
-    resultStorage.append(31)
-    
-    XCTAssertEqual(testArray.storage.count, 13)
-    XCTAssertEqual(newArr.storage, resultStorage)
-    
-  }
-  
-  func testBitwiseAND(){
-    var testArray = BitArray()
-    var testArray2 = BitArray()
-    
-    for i in 0...limit {
-      testArray.append((i%2 == 0))
-    }
-    
-    for i in 0...limit {
-      testArray2.append((i%2 != 0))
-    }
-    
-    
-    let newArr = testArray.bitwiseAnd(with: testArray2)
-    
-    var resultStorage: [UInt8] = []
-    
-    for _ in 0..<12 {
-      resultStorage.append(0)
-    }
-    resultStorage.append(0)
-    
-    XCTAssertEqual(testArray.storage.count, 13)
-    XCTAssertEqual(newArr.storage, resultStorage)
-    
-  }
-  
-  func testBitwiseXOR(){
-    var testArray = BitArray()
-    var testArray2 = BitArray()
-    
-    for i in 0...limit {
-      testArray.append((i%2 == 0))
-    }
-    
-    for i in 0...limit {
-      testArray2.append((i%2 != 0))
-    }
-    
-    
-    let newArr = testArray.bitwiseXor(with: testArray2)
-    
-    var resultStorage: [UInt8] = []
-    
-    for _ in 0..<12 {
-      resultStorage.append(BitArray.UNIT.max)
-    }
-    resultStorage.append(31)
-    
-    XCTAssertEqual(testArray.storage.count, 13)
-    XCTAssertEqual(newArr.storage, resultStorage)
-    
-  }
-  
-  func testToggle() { // toggle function already works?
-    var testArray = BitArray()
-    var value = Bool.random()
-    
-    
-    for _ in 0...limit {
-      testArray.append(value)
-      value = Bool.random()
-    }
-    
-    var copyArray = testArray
-    
-    // will take each value and toggle it back and forth 4 times, making sure the rest of the array is unaffected each time
-    for i in 0...limit {
-      copyArray = testArray
-      value = testArray[i]
-      
-      testArray[i].toggle()
-      XCTAssertEqual(!value, testArray[i])
-      testArray[i].toggle()
-      XCTAssertEqual(value, testArray[i])
-      XCTAssertEqual(copyArray, testArray)
-    }
-    
-  }
-  
 }
